Оптимизация кодогенерации
