+++ lib/app/data/models/base/description_base_model.codegen.dart
@@ -1,9 +1,9 @@
 import 'package:json_annotation/json_annotation.dart';
 
 import '../json_parsers/description_parser.dart';
-import 'reference_base_model.dart';
+import 'reference_base_model.codegen.dart';
 
-part 'description_base_model.g.dart';
+part 'description_base_model.codegen.g.dart';
 
 /// большое кол-во сущностей включат в себя описани
 @JsonSerializable()
