+++ lib/app/data/models/base/reference_base_model.codegen.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/foundation.dart';
 import 'package:json_annotation/json_annotation.dart';
 
-part 'reference_base_model.g.dart';
+part 'reference_base_model.codegen.g.dart';
 
 /// базовая сущность всего dnd 5e API
 @JsonSerializable()
