+++ lib/app/network/dnd5e_api.dart
@@ -1,9 +1,9 @@
 import 'package:dio/dio.dart';
 import 'package:injectable/injectable.dart';
 
-import '../data/models/base/reference_base_model.dart';
+import '../data/models/base/reference_base_model.codegen.dart';
 import '../data/models/json_parsers/reference_model_converter.dart';
-import '../data/models/resource_list_model.dart';
+import '../data/models/resource_list_model.codegen.dart';
 import 'dio_container.dart';
 
 @Injectable()
