+++ lib/app/presentation/router/app_router.dart
@@ -10,7 +10,7 @@ import '../../../features/library/presentation/pages/library_container_page.dart
 import '../../../features/library/presentation/pages/library_item_page.dart';
 import '../../../features/library/presentation/pages/library_page.dart';
 import '../../../features/settings/presentation/settings_page.dart';
-import '../../data/models/base/reference_base_model.dart';
+import '../../data/models/base/reference_base_model.codegen.dart';
 
 part 'app_router.gr.dart';
