+++ lib/features/library/data/repositories/library_category_repo_impl.dart
@@ -1,6 +1,6 @@
 import 'package:injectable/injectable.dart';
 
-import '../../../../app/data/models/base/reference_base_model.dart';
+import '../../../../app/data/models/base/reference_base_model.codegen.dart';
 import '../../../../app/network/dnd5e_api.dart';
 import '../../domain/entities/library_category_entity.dart';
 import '../../domain/repositories/library_repo.dart';
