+++ lib/features/library/domain/repositories/library_repo.dart
@@ -1,4 +1,4 @@
-import '../../../../app/data/models/base/reference_base_model.dart';
+import '../../../../app/data/models/base/reference_base_model.codegen.dart';
 import '../entities/library_category_entity.dart';
 
 abstract class LibraryRepo {
