+++ lib/features/library/domain/state/library_category/library_category_cubit.dart
@@ -1,7 +1,7 @@
 import 'package:flutter/widgets.dart';
 import 'package:flutter_bloc/flutter_bloc.dart';
 
-import '../../../../../app/data/models/base/reference_base_model.dart';
+import '../../../../../app/data/models/base/reference_base_model.codegen.dart';
 import '../../../../../app/domain/cubit/gin_cubit.dart';
 import '../../../../../app/domain/model/async_field.dart';
 import '../../entities/library_category_entity.dart';
