+++ lib/features/library/domain/state/library_item/library_item_state.codegen.dart
@@ -1,6 +1,6 @@
 import 'package:freezed_annotation/freezed_annotation.dart';
 
-import '../../../../../app/data/models/base/reference_base_model.dart';
+import '../../../../../app/data/models/base/reference_base_model.codegen.dart';
 import '../../../../../app/domain/model/async_field.dart';
 
 part 'library_item_state.codegen.freezed.dart';
