+++ lib/features/library/presentation/components/category_item_card.dart
@@ -1,7 +1,7 @@
 import 'package:auto_route/auto_route.dart';
 import 'package:flutter/material.dart';
 
-import '../../../../app/data/models/base/reference_base_model.dart';
+import '../../../../app/data/models/base/reference_base_model.codegen.dart';
 import '../../../../app/presentation/router/app_router.dart';
 import '../../domain/state/library_category/library_category_cubit.dart';
