+++ lib/features/library_character_data/data/models/alignment_model.codegen.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../../app/data/models/base/description_base_model.dart';
+import '../../../../app/data/models/base/description_base_model.codegen.dart';
 import '../../../../app/data/models/json_parsers/description_parser.dart';
 
 part 'alignment_model.codegen.g.dart';
@@ -9,6 +9,7 @@ part 'alignment_model.codegen.g.dart';
 class AlignmentModel extends DescribedBaseModel {
   const AlignmentModel({
     required this.abbreviation,
+    required this.abbreviation1,
     required super.url,
     required super.index,
     required super.name,
@@ -19,6 +20,7 @@ class AlignmentModel extends DescribedBaseModel {
       _$AlignmentModelFromJson(json);
 
   final String abbreviation;
+  final String abbreviation1;
 
   @override
   Map<String, dynamic> toJson() => _$AlignmentModelToJson(this);
