+++ lib/features/library_character_data/data/models/language_model.codegen.dart
@@ -1,6 +1,6 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../../app/data/models/base/description_base_model.dart';
+import '../../../../app/data/models/base/description_base_model.codegen.dart';
 import '../../../../app/data/models/json_parsers/description_parser.dart';
 
 part 'language_model.codegen.g.dart';
