+++ lib/features/library_character_data/data/models/skill_model.codegen.dart
@@ -1,7 +1,7 @@
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../../app/data/models/base/description_base_model.dart';
-import '../../../../app/data/models/base/reference_base_model.dart';
+import '../../../../app/data/models/base/description_base_model.codegen.dart';
+import '../../../../app/data/models/base/reference_base_model.codegen.dart';
 import '../../../../app/data/models/json_parsers/description_parser.dart';
 
 part 'skill_model.codegen.g.dart';
