+++ lib/features/library_game_mechanics/data/models/damage_type_model.codegen.dart
@@ -1,10 +1,10 @@
 import 'package:flutter/foundation.dart';
 import 'package:json_annotation/json_annotation.dart';
 
-import '../../../../app/data/models/base/description_base_model.dart';
+import '../../../../app/data/models/base/description_base_model.codegen.dart';
 import '../../../../app/data/models/json_parsers/description_parser.dart';
 
-part 'damage_type_model.g.dart';
+part 'damage_type_model.codegen.g.dart';
 
 @JsonSerializable(fieldRename: FieldRename.snake)
 @immutable
