+++ pubspec.yaml
@@ -30,6 +30,7 @@ dependencies:
   freezed_annotation: ^2.0.3
   auto_route: ^4.0.1
   provider: ^6.0.3
+  build_config: ^1.0.0
 
 dev_dependencies:
   auto_route_generator: ^4.0.0
